+++ packages/core/src/awsService/appBuilder/explorer/nodes/deployedNode.ts
@@ -16,7 +16,7 @@ import { defaultPartition } from '../../../../shared/regions/regionProvider'
 import { Lambda, APIGateway } from 'aws-sdk'
 import { LambdaNode } from '../../../../lambda/explorer/lambdaNodes'
 import { LambdaFunctionNode } from '../../../../lambda/explorer/lambdaFunctionNode'
-import { DefaultS3Client, DefaultBucket } from '../../../../shared/clients/s3Client'
+import { DefaultS3Client, DefaultBucket } from '../../../../shared/clients/s3'
 import { S3Node } from '../../../../awsService/s3/explorer/s3Nodes'
 import { S3BucketNode } from '../../../../awsService/s3/explorer/s3BucketNode'
 import { ApiGatewayNode } from '../../../../awsService/apigateway/explorer/apiGatewayNodes'
