+++ packages/core/src/awsService/s3/commands/downloadFileAs.ts
@@ -12,7 +12,7 @@ import { readablePath } from '../util'
 import { progressReporter } from '../progressReporter'
 import { localize } from '../../../shared/utilities/vsCodeUtils'
 import { showOutputMessage } from '../../../shared/utilities/messages'
-import { DefaultS3Client, S3Client } from '../../../shared/clients/s3Client'
+import { DefaultS3Client, S3Client } from '../../../shared/clients/s3'
 import { Timeout, CancellationError } from '../../../shared/utilities/timeoutUtils'
 import { ToolkitError } from '../../../shared/errors'
 import { streamToBuffer, streamToFile } from '../../../shared/utilities/streamUtilities'
