+++ packages/core/src/awsService/s3/commands/uploadFile.ts
@@ -15,7 +15,7 @@ import { localize } from '../../../shared/utilities/vsCodeUtils'
 import { showOutputMessage } from '../../../shared/utilities/messages'
 import { createQuickPick, promptUser, verifySinglePickerOutput } from '../../../shared/ui/picker'
 import { addCodiconToString } from '../../../shared/utilities/textUtilities'
-import { Bucket, Folder, S3Client } from '../../../shared/clients/s3Client'
+import { Bucket, Folder, S3Client } from '../../../shared/clients/s3'
 import { createBucketCommand } from './createBucket'
 import { S3BucketNode } from '../explorer/s3BucketNode'
 import { S3FolderNode } from '../explorer/s3FolderNode'
