+++ packages/core/src/awsService/s3/explorer/s3BucketNode.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from 'vscode'
 import { ChildNodePage } from '../../../awsexplorer/childNodeLoader'
-import { Bucket, CreateFolderRequest, CreateFolderResponse, S3Client } from '../../../shared/clients/s3Client'
+import { Bucket, CreateFolderRequest, CreateFolderResponse, S3Client } from '../../../shared/clients/s3'
 
 import { AWSResourceNode } from '../../../shared/treeview/nodes/awsResourceNode'
 import { AWSTreeNodeBase } from '../../../shared/treeview/nodes/awsTreeNodeBase'
