+++ packages/core/src/awsexplorer/regionNode.ts
@@ -26,7 +26,7 @@ import { DefaultAppRunnerClient } from '../shared/clients/apprunnerClient'
 import { DefaultEcrClient } from '../shared/clients/ecrClient'
 import { DefaultRedshiftClient } from '../shared/clients/redshiftClient'
 import { DefaultIotClient } from '../shared/clients/iotClient'
-import { DefaultS3Client } from '../shared/clients/s3Client'
+import { DefaultS3Client } from '../shared/clients/s3'
 import { DefaultSchemaClient } from '../shared/clients/schemaClient'
 import { getEcsRootNode } from '../awsService/ecs/model'
 import { compareTreeItems, TreeShim } from '../shared/treeview/utils'
