+++ packages/core/src/dynamicResources/explorer/nodes/resourceTypeNode.ts
@@ -17,7 +17,7 @@ import { ResourceNode } from './resourceNode'
 import { Result } from '../../../shared/telemetry/telemetry'
 import { CloudControl } from 'aws-sdk'
 import { ResourceTypeMetadata } from '../../model/resources'
-import { DefaultS3Client } from '../../../shared/clients/s3Client'
+import { DefaultS3Client } from '../../../shared/clients/s3'
 import { telemetry } from '../../../shared/telemetry/telemetry'
 
 export const contextValueResourceOperations: any = {
