+++ packages/core/src/shared/sam/deploy.ts
@@ -9,7 +9,7 @@ import { TreeNode, isTreeNode } from '../treeview/resourceTreeDataProvider'
 import globals from '../../shared/extensionGlobals'
 import { ToolkitError } from '../../shared/errors'
 import { DefaultCloudFormationClient } from '../clients/cloudFormationClient'
-import { DefaultS3Client } from '../clients/s3Client'
+import { DefaultS3Client } from '../clients/s3'
 import { samDeployUrl } from '../constants'
 import { getSpawnEnv } from '../env/resolveEnv'
 import { CloudFormationTemplateRegistry } from '../fs/templateRegistry'
