+++ packages/core/src/shared/sam/sync.ts
@@ -8,7 +8,7 @@ import globals from '../extensionGlobals'
 import * as vscode from 'vscode'
 import * as path from 'path'
 import * as localizedText from '../localizedText'
-import { DefaultS3Client } from '../clients/s3Client'
+import { DefaultS3Client } from '../clients/s3'
 import { DataQuickPickItem, createMultiPick, createQuickPick } from '../ui/pickerPrompter'
 import { DefaultCloudFormationClient } from '../clients/cloudFormationClient'
 import * as CloudFormation from '../cloudformation/cloudformation'
