+++ packages/core/src/test/awsService/accessanalyzer/iamPolicyChecks.test.ts
@@ -13,8 +13,8 @@ import {
 } from '../../../awsService/accessanalyzer/vue/iamPolicyChecks'
 import { globals } from '../../../shared'
 import { AccessAnalyzer, Config } from 'aws-sdk'
-import * as s3Client from '../../../shared/clients/s3Client'
-import { DefaultS3Client } from '../../../shared/clients/s3Client'
+import * as s3Client from '../../../shared/clients/s3'
+import { DefaultS3Client } from '../../../shared/clients/s3'
 import * as iamPolicyChecks from '../../../awsService/accessanalyzer/vue/iamPolicyChecks'
 import * as vscode from 'vscode'
 import { IamPolicyChecksConstants } from '../../../awsService/accessanalyzer/vue/constants'
