+++ packages/core/src/test/awsService/s3/explorer/s3FolderNode.test.ts
@@ -7,7 +7,7 @@ import assert from 'assert'
 import { MoreResultsNode } from '../../../../awsexplorer/moreResultsNode'
 import { S3FileNode } from '../../../../awsService/s3/explorer/s3FileNode'
 import { S3FolderNode } from '../../../../awsService/s3/explorer/s3FolderNode'
-import { S3Client, File, Folder, Bucket } from '../../../../shared/clients/s3Client'
+import { S3Client, File, Folder, Bucket } from '../../../../shared/clients/s3'
 import { AWSTreeNodeBase } from '../../../../shared/treeview/nodes/awsTreeNodeBase'
 import { LoadMoreNode } from '../../../../shared/treeview/nodes/loadMoreNode'
 import { TestSettings } from '../../../utilities/testSettingsConfiguration'
