+++ packages/core/src/test/awsService/s3/util/fileViewerManager.test.ts
@@ -7,7 +7,7 @@ import assert from 'assert'
 import { ManagedUpload } from 'aws-sdk/clients/s3'
 import * as vscode from 'vscode'
 import { S3FileProvider, S3FileViewerManager } from '../../../../awsService/s3/fileViewerManager'
-import { DefaultBucket, DefaultS3Client, File, toFile } from '../../../../shared/clients/s3Client'
+import { DefaultBucket, DefaultS3Client, File, toFile } from '../../../../shared/clients/s3'
 import globals from '../../../../shared/extensionGlobals'
 import { VirtualFileSystem } from '../../../../shared/virtualFilesystem'
 import { bufferToStream } from '../../../../shared/utilities/streamUtilities'
