+++ packages/core/src/test/shared/clients/defaultS3Client.test.ts
@@ -8,8 +8,8 @@ import { AWSError, Request, S3 } from 'aws-sdk'
 import { DeleteObjectsRequest, ListObjectVersionsOutput, ListObjectVersionsRequest } from 'aws-sdk/clients/s3'
 import { FileStreams } from '../../../shared/utilities/streamUtilities'
 import * as vscode from 'vscode'
-import { DefaultBucket, DefaultFolder, DefaultS3Client, toFile } from '../../../shared/clients/s3Client'
-import { DEFAULT_DELIMITER, DEFAULT_MAX_KEYS } from '../../../shared/clients/s3Client'
+import { DefaultBucket, DefaultFolder, DefaultS3Client, toFile } from '../../../shared/clients/s3'
+import { DEFAULT_DELIMITER, DEFAULT_MAX_KEYS } from '../../../shared/clients/s3'
 import { FakeFileStreams } from './fakeFileStreams'
 import globals from '../../../shared/extensionGlobals'
 import sinon from 'sinon'
