+++ packages/core/src/test/shared/sam/deploy.test.ts
@@ -18,9 +18,9 @@ import { intoCollection } from '../../../shared/utilities/collectionUtils'
 import { clickBackButton, createPromptHandler, PrompterTester } from '../wizards/prompterTester'
 import { RegionNode } from '../../../awsexplorer/regionNode'
 import { createTestRegionProvider } from '../regions/testUtil'
-import { DefaultS3Client } from '../../../shared/clients/s3Client'
+import { DefaultS3Client } from '../../../shared/clients/s3'
 import * as CloudFormationClientModule from '../../../shared/clients/cloudFormationClient'
-import * as S3ClientModule from '../../../shared/clients/s3Client'
+import * as S3ClientModule from '../../../shared/clients/s3'
 import * as ProcessUtilsModule from '../../../shared/utilities/processUtils'
 import * as ProcessTerminalModule from '../../../shared/sam/processTerminal'
 import * as ResolveEnvModule from '../../../shared/env/resolveEnv'
