+++ packages/core/src/test/shared/sam/sync.test.ts
@@ -6,7 +6,7 @@
 import * as vscode from 'vscode'
 import * as SamUtilsModule from '../../../shared/sam/utils'
 import * as ProcessTerminalUtils from '../../../shared/sam/processTerminal'
-import * as S3ClientModule from '../../../shared/clients/s3Client'
+import * as S3ClientModule from '../../../shared/clients/s3'
 import * as SamConfigModule from '../../../shared/sam/config'
 import * as ResolveEnvModule from '../../../shared/env/resolveEnv'
 import * as ProcessUtilsModule from '../../../shared/utilities/processUtils'
@@ -41,7 +41,7 @@ import fs from '../../../shared/fs/fs'
 import { createMultiPick, DataQuickPickItem } from '../../../shared/ui/pickerPrompter'
 import sinon from 'sinon'
 import { getTestWindow } from '../vscode/window'
-import { DefaultS3Client } from '../../../shared/clients/s3Client'
+import { DefaultS3Client } from '../../../shared/clients/s3'
 import { RequiredProps } from '../../../shared/utilities/tsUtils'
 import S3 from 'aws-sdk/clients/s3'
 import { DefaultCloudFormationClient } from '../../../shared/clients/cloudFormationClient'
